+++ emp-tool/emp-tool/utils/ThreadPool.h
@@ -37,7 +37,7 @@ freely, subject to the following restrictions:
 
 class ThreadPool {
 public:
-    ThreadPool(uint64_t);
+    ThreadPool(size_t);
     template<class F, class... Args>
     auto enqueue(F&& f, Args&&... args) 
         -> std::future<typename std::result_of<F(Args...)>::type>;
@@ -59,10 +59,10 @@ int inline ThreadPool::size() const {
 	return workers.size();
 }
 // the constructor just launches some amount of workers
-inline ThreadPool::ThreadPool(uint64_t threads)
+inline ThreadPool::ThreadPool(size_t threads)
     :   stop(false)
 {
-    for(uint64_t i = 0;i<threads;++i)
+    for(size_t i = 0;i<threads;++i)
         workers.emplace_back(
             [this]
             {
